+++ src/aks-preview/azext_aks_preview/_params.py
@@ -241,8 +241,8 @@ from azext_aks_preview._validators import (
     validate_duration_hours,
     validate_vm_set_type,
     validate_vnet_subnet_id,
-    validate_system_node_vnet_subnet_id,
-    validate_node_vnet_subnet_id,
+    validate_system_node_subnet_id,
+    validate_node_subnet_id,
     validate_force_upgrade_disable_and_enable_parameters,
     validate_azure_service_mesh_revision,
     validate_artifact_streaming,
@@ -1312,15 +1312,15 @@ def load_arguments(self, _):
         c.argument("enable_hosted_system", action="store_true", is_preview=True)
         c.argument("disable_hosted_system", action="store_true", is_preview=True)
         c.argument(
-            "system_node_vnet_subnet_id",
-            options_list=["--system-node-vnet-subnet-id", "--sys-node-subnet-id"],
-            validator=validate_system_node_vnet_subnet_id,
+            "system_node_subnet_id",
+            options_list=["--system-node-subnet-id", "--sys-node-subnet-id"],
+            validator=validate_system_node_subnet_id,
             is_preview=True,
         )
         c.argument(
-            "node_vnet_subnet_id",
-            options_list=["--node-vnet-subnet-id"],
-            validator=validate_node_vnet_subnet_id,
+            "node_subnet_id",
+            options_list=["--node-subnet-id"],
+            validator=validate_node_subnet_id,
             is_preview=True,
         )
         c.argument(